+++ lib/sequent/core/stream_record.rb
@@ -4,10 +4,26 @@ require 'active_record'
 
 module Sequent
   module Core
-    EventStream = Data.define(:aggregate_type, :aggregate_id, :events_partition_key, :snapshot_outdated_at) do
-      def initialize(aggregate_type:, aggregate_id:, events_partition_key: '', snapshot_outdated_at: nil)
+    EventStream = Data.define(
+      :aggregate_type,
+      :aggregate_id,
+      :events_partition_key,
+      :snapshot_outdated_at,
+      :unique_keys,
+    ) do
+      def initialize(aggregate_type:, aggregate_id:, events_partition_key: '', snapshot_outdated_at: nil,
+                     unique_keys: {})
         super
       end
+    end
+
+    class AggregateUniqueKey < Sequent::ApplicationRecord
+      self.primary_key = %i[aggregate_id scope]
+      self.table_name = 'aggregate_unique_keys'
+
+      validates_presence_of :aggregate_id, :scope, :key
+
+      belongs_to :stream_record, foreign_key: :aggregate_id, primary_key: :aggregate_id
     end
 
     class StreamRecord < Sequent::ApplicationRecord
@@ -18,12 +34,14 @@ module Sequent
       validates_presence_of :aggregate_type, :aggregate_id
 
       has_many :event_records, foreign_key: :aggregate_id, primary_key: :aggregate_id
+      has_many :aggregate_unique_keys, foreign_key: :aggregate_id, primary_key: :aggregate_id
 
       def event_stream
         EventStream.new(
           aggregate_type:,
           aggregate_id:,
           events_partition_key:,
+          unique_keys: aggregate_unique_keys.to_h { |key| [key.scope.to_sym, key.key] },
         )
       end